Rough draft of DataFile upgrade routine refactor.
Split several large post 1.2.2 upgrade routines out into functors
Included a template for making more upgrade routine functors
