+++ include/UpgradeExtendedNoteRange.h
