+++ include/UpgradeSampleAndHold.h
@@ -1,0 +1,48 @@
+/*
+ * UpgradeSampleAndHold.h - functor for upgrading data files for new sample
+ *     and hold functionality, so freq > 0 won't change the project.
+ *
+ * This file is part of LMMS - https://lmms.io
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public
+ * License as published by the Free Software Foundation; either
+ * version 2 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public
+ * License along with this program (see COPYING); if not, write to the
+ * Free Software Foundation, Inc., 51 Franklin Street, Fifth Floor,
+ * Boston, MA 02110-1301 USA.
+ *
+ */
+
+#pragma once
+
+#ifndef LMMS_UPGRADESAMPLEANDHOLD_H
+#define LMMS_UPGRADESAMPLEANDHOLD_H
+
+#include "DataFileUpgrade.h"
+
+
+namespace lmms
+{
+
+// upgrade functor for 
+class UpgradeSampleAndHold : public DataFileUpgrade
+{
+public:
+	UpgradeSampleAndHold(QDomDocument& document) : DataFileUpgrade(document) {}
+
+	virtual void upgrade() override;
+};
+
+
+}
+
+#endif // LMMS_UPGRADESAMPLEANDHOLD_H
+
